Update saphon_json_automation.htm
This updated added a partially working auto-resizer for the text area. Because we are loading JSON files and importing its content into the text area, an event is necessary to trigger expand_height in json editor. The code I added triggers this in the main page. Even though my method should apply to all text areas, the ones hidden under other tabs don't seem to be affected. I don't know how to penetrate the barrier of these tabs.

Also, after the code that Ron pushed, the "restore to default" button no longer worked. Specifically, the modal does not show up, and even when I forced it to be visible with CSS, the button "YES" does nothing. This is a work in progress.